requirements.txt
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
ply==3.11
pytest==8.1.1
svg.py==1.5.0
